CALL: More compact encoding of call header - FIX: use "short" for both service and methodId - FIX: special encoding to combine service changed and param size

+++ call/src/main/java/org/kari/call/ServiceName.java
@@ -8,6 +8,6 @@ import java.lang.annotation.Target;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface ServiceName {
-    int id();
+    short id();
     String name();
 }
